requirements.txt
altgraph==0.17.1
appdirs==1.4.4
asgiref==3.7.2
auto-py-to-exe==2.13.0
bottle==0.12.25
bottle-websocket==0.2.9
CacheControl==0.13.1
certifi==2024.6.2
chardet==5.2.0
colorama==0.4.6
contextlib2==21.6.0
distlib==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.11
django-heroku==0.3.1
Eel==0.12.4
future==0.18.3
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==21.2.0
html5lib==1.1
idna==3.7
ipaddr==2.2.0
lockfile==0.12.2
msgpack==1.0.7
mysql-connector-python==8.3.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pep517==0.13.0
Pillow==10.3.0
progress==1.6
protobuf==5.26.1
psycopg2==2.9.9
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
pytoml==0.1.21
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
whichcraft==0.6.1
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.3
djangorestframework==3.14.0
